refactor: scenes/ディレクトリ作成とBootScene移行（TASK-0094）
BootSceneをsrc/presentation/scenes/からsrc/scenes/に移行。
presentation側は再エクスポートシムに変更して後方互換性を維持。
src/scenes/index.tsを作成してBootSceneをエクスポート。
