+++ atelier-guild-rank/src/scenes/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Scenes Module - Phaserシーン公開API
+ *
+ * TASK-0094: scenes/ディレクトリ作成とBootScene移行
+ *
+ * @description
+ * Phaserシーンの公開エクスポート。
+ * 各シーンは段階的にsrc/scenes/に移行される。
+ */
+
+export { BootScene } from './BootScene';
